+++ lib/happn.rb
@@ -54,7 +54,6 @@ module Happn
     config.on_error                   = nil
     config.bunny_options              = {}
     config.management_options         = {}
-    config.on_error                   = nil
   end
 
   private